// src/app/core/api/todo/model/createTodoV1.ts
/**
 * Todo (development)
 * API for the TODO
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CreateTodoV1 { 
    title?: string;
    completed?: boolean;
    priorityId?: number;
}